[FE #12-1-2] fix : Form 버그데이 피드백 반영 및 UX 고도화 및 검증 기능 강화 (#354)
* fix: PDF 다운로드 되는 동안 로딩 표시, 입력 타입 숫자->텍스트에 통합

* fix: 폼 최대 입력 라인 추가

* feat: form 날짜 및 숫자에 대한 validation 추가

* feat: selector 중복 값 허용 x 기능 추가

* feat: 국적 필드 선택 해제 기능

* fix: -1값일 경우 세지 않도록 수정

* fix: 여러 줄 있을 때 줄 삭제 전파 수정

* fix: selector input으로 완성

* feat: 검색 결과 없음 대응

* fix: isDone이어도 로컬 기록 불러오기 활성화

* feat: 수임 수락한 외국인은 비자 신청서 readonly

* fix: readOnly 기능 완성

* fix: codeRabbit 제안 사항 적용

* fix: formSelector key가 string이어도 삭제 가능

+++ FE/src/pages/Documents/Component/EditDocumentWidget.tsx
@@ -7,10 +7,11 @@ import ProgressStepWidget from "../../../components/form/ProgressStepWidget";
 import type { FormSection } from "../../../types/formType";
 import { useEffect, useState } from "react";
 import FloatingChatModal from "./FloatingChatModal";
-import { useGeneratePdf } from "./hooks/useGeneratePdf";
+import { useGeneratePdf } from "../hooks/useGeneratePdf";
 import ConfirmToExportModal from "./ConfirmToExportModal";
 import { usePatchFormStatusMutation } from "../../../api/mutations/usePatchFormStatusMutation";
 import { useNavigate } from "react-router-dom";
+import GeneratingModal from "./GeneratingModal";
 
 const EditDocumentWidget = ({
   editDocumentData,
@@ -22,30 +23,33 @@ const EditDocumentWidget = ({
   goTop,
   documentId,
   chatRoomId,
+  readOnly = false,
 }: EditDocumentWidgetProps) => {
   const { t } = useTranslation(["pages"]);
   const filledFormData = useWatch();
   const navigate = useNavigate();
-  const { previewPdf, downloadPdf } = useGeneratePdf();
+  const { previewPdf, downloadPdf, generating } = useGeneratePdf();
   const [confirmModalOn, setConfirmModalOn] = useState(false);
   const patchStatus = usePatchFormStatusMutation(() => setConfirmModalOn(false));
   const [isChatOpen, setIsChatOpen] = useState<boolean>(false);
 
   useEffect(() => {
-    const formValues = Object.values(filledFormData).slice(0, 9);
-    const data = {
-      updatedAt: new Date(),
-      sections: formValues.map((section, i) => ({
-        sectionId: i + 1,
+    if (!readOnly) {
+      const formValues = Object.values(filledFormData).slice(0, 9);
+      const data = {
+        updatedAt: new Date(),
+        sections: formValues.map((section, i) => ({
+          sectionId: i + 1,
 
-        sectionData: section.sectionData.map((field: Record<string, any>) => ({
-          ...field,
-          values: field.values && Object.values(field.values),
+          sectionData: section.sectionData.map((field: Record<string, any>) => ({
+            ...field,
+            values: field.values && Object.values(field.values),
+          })),
         })),
-      })),
-    };
-    window.localStorage.setItem(documentId, JSON.stringify(data));
-  }, [filledFormData, documentId]);
+      };
+      window.localStorage.setItem(documentId, JSON.stringify(data));
+    }
+  }, [readOnly, filledFormData, documentId]);
 
   const handlePreviewPdf = () => previewPdf(filledFormData, imageUrl);
   const handleDownloadPdf = () => {
@@ -78,6 +82,7 @@ const EditDocumentWidget = ({
   return (
     <div className="w-fit ml-4 left-0 mt-17 flex flex-row">
       {confirmModalOn && <ConfirmToExportModal onCancel={onCancel} onConfirm={onConfirm} />}
+      {generating && <GeneratingModal />}
       <div className="flex flex-col w-92 gap-5 ">
         <Button variant="primary" className="drop-shadow-[0_0_7px_#6860A040]" type="submit">
           {t("documents.save")}
@@ -124,4 +129,5 @@ type EditDocumentWidgetProps = {
   goTop: () => void;
   documentId: string;
   chatRoomId: number | null;
+  readOnly: boolean;
 };
